### 2023-06-21 #5578 @editorialbot
[PRE REVIEW]: archeoViz: an R package for the Visualisation, Exploration, and Web Communication of Archaeological Spatial Data
**Submitting author:** @sebastien-plutniak (<a href="http://orcid.org/0000-0002-6674-3806">Sébastien Plutniak</a>)
**Repository:** https://github.com/sebastien-plutniak/archeoviz
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.2
