### 2023-06-21 #5578 @sebastien-plutniak
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.14434/sdh.v2i2.24446 is OK
> - 10.1109/DigitalHeritage.2015.7419505 is OK
> - 10.5281/zenodo.7938998 is OK
> - 10.5334/jcaa.81 is OK
> - 10.1201/9781315370279 is OK
> - 10.1007/978-3-319-24277-4 is OK
> - 10.5281/zenodo.7674698 is OK
>
> MISSING DOIs
>
> - 10.1201/9780429447273 may be a valid DOI for title: Interactive Web-Based Data Visualization with R, plotly, and shiny
>
> INVALID DOIs
>
> - None
> ```

Thanks, I add the missing DOI
