### 2023-06-21 91e4353
add JOSS badge and update .bib
