+++ paper.bib
@@ -7,13 +7,24 @@
   Publisher                = {Chapman \& Hall/CRC},
   Series                   = {The R series},
   Url                      = {https://plotly-r.com},
-
+  Doi                      = {10.1201/9780429447273},
   Address                  = {Boca Raton (Fla.)},
-  Keywords                 = {none},
   Langid                   = {american}
 }
 
 
+
+
+@Article{OrtegaalvaradoEtal2022,
+  Title                    = {Integrated and Interactive 4D System for Archaeological Stratigraphy},
+  Author                   = {Ortega-Alvarado, Lidia M., García-Fernández,   Ángel-Luis and Conde-Rodríguez, Francisco and  Jurado-Rodríguez, Juan M.},
+  Year                     = {2022},
+  Doi                      = {10.1007/s12520-022-01667-3},
+  Number                   = {203},
+  Volume                   = {14},
+  Journal                  = {Archaeological and Anthropological Sciences},
+  Langid                   = {american}
+}
  
  @article{VurpillotEtal2018,
   title={Aspectus: A Flexible Collaboration Tool for Multimodal and Multiscalar 3D Data Exploitation},
@@ -24,7 +35,6 @@
   number={2},
   pages={150--165},
   year={2018},
-  Keywords                 = {archeo},
   Langid                   = {american}
 }
 
@@ -38,7 +48,6 @@
   volume={2},
   pages={251-258},
   doi={10.1109/DigitalHeritage.2015.7419505},
-  Keywords                 = {archeo},
   Langid                   = {american}
   }
 
@@ -99,7 +108,6 @@
     isbn = {978-1315370279},
     edition = {2},
     publisher = {Chapman and Hall/CRC},
-  Keywords                 = {none},
   Langid                   = {american}
   }
 
@@ -109,7 +117,6 @@
     year = {2023},
     note = {R package version 0.7.2},
     url = {https://CRAN.R-project.org/package=cxhull},
-  Keywords                 = {none},
   Langid                   = {american}
   }
   
@@ -119,7 +126,6 @@
     year = {2023},
     note = {R package version 1.6.2},
     url = {https://CRAN.R-project.org/package=htmlwidgets},
-  Keywords                 = {none},
   Langid                   = {american}
   }
 
@@ -136,7 +142,6 @@
   Url                      = {https://ggplot2.tidyverse.org},
 
   Address                  = {New York},
-  Keywords                 = {none},
   Langid                   = {american}
 }
 
@@ -149,7 +154,6 @@
   pubstate                   = {forthcoming},
 
   Journal                  = {PCI Archaeology},
-  Keywords                 = {plutniak, archeo, notyet, hceres-acl},
   Langid                   = {american}
 }
 @Manual{Rcoreteam2023,
@@ -160,7 +164,6 @@
   Url                      = {https://www.R-project.org/},
 
   Address                  = {Vienna, Austria},
-  Keywords                 = {software},
   Langid                   = {american}
 }
 
@@ -181,7 +184,6 @@
   Pages                    = {70--74},
   Volume                   = {120},
   Journal                  = {Bulletin de la Société préhistorique française},
-  Keywords                 = {plutniak, archeo, other, hceres-ascl},
   Langid                   = {french}
 }
 
@@ -195,6 +197,5 @@
   Pagetotal                = {135},
   note = {Version 2.2},
    url = {http://legacy.ariadne-infrastructure.eu/wp-content/uploads/2019/01/ARIADNE_D2.2_Second_report_on_users_needs.pdf},
-  Keywords                 = {archeo},
   Langid                   = {american}
 }
