+++ paper.bib
@@ -12,7 +12,17 @@
   Langid                   = {american}
 }
 
-
+@Article{Perrin2019,
+  Title                    = {BDA : une Base de Données Archéologique collaborative en ligne},
+  Author                   = {Perrin, Thomas},
+  Year                     = {1960},
+  Number                   = {1},
+  Pages                    = {159‑162},
+  doi                      = {10.3406/bspf.2019.14988},
+  Volume                   = {116},
+  Journal                  = {Bulletin de la Société préhistorique française},
+  Langid                   = {french}
+}
 
 
 @Article{OrtegaalvaradoEtal2022,
