### 2023-06-26 4edce08
Update german interface
Adding German translation (beta) to "load_interface_terms_utf8.R".
