+++ R/load_interface_terms_utf8.R
@@ -23,6 +23,7 @@
                 L'interface est disponible en anglais, français, italien et portugais.
                 </p>
                 Le package, sous licence libre, est publié sur le <a href=https://cran.r-project.org/package=archeoViz target=_blank>CRAN</a>. Son code source est aussi disponible sur <a href=https://github.com/sebastien-plutniak/archeoviz target=_blank>github</a>."
+  
   # : en ----
   welcome.en <- "
                 <h1>Welcome to <i>archeoViz</i>!</h1>
@@ -45,6 +46,7 @@
                 The interface is available in English, French, Italian and Portuguese.
                 </p>
                 The package is published under a free license on the <a href=https://cran.r-project.org/package=archeoViz target=_blank>CRAN</a>. Its source code is also available on  <a href=https://github.com/sebastien-plutniak/archeoviz target=_blank>github</a>."
+  
   # : pt ----
   welcome.pt <- "<h1>Bem-vindo ao <i>archeoViz</i>!</h1>
                 <p><i>archeoViz</i> é um aplicativo dedicado à arqueologia. Ele permite <b>visualizar</b>, <b>explorar</b> interativamente, e   
@@ -70,6 +72,37 @@
                 Distribuito come pacchetto R, <i>archeoViz</i> può essere utilizzato localmente o distribuito su un server, caricando i dati tramite l’interfaccia o lanciando l’applicazione con un insieme di dati specifico. L’interfaccia è disponibile in inglese, francese, italiano e portoghese.
                 </p>
                 Il pacchetto è open source e pubblicato su il <a href=https://cran.r-project.org/package=archeoViz target=_blank>CRAN</a> Il suo codice sorgente è disponibile anche su  <a href=https://github.com/sebastien-plutniak/archeoviz target=_blank>github</a>."
+   
+   # : de ----
+   welcome.de <- "<h1>Willkommen bei <i>archeoViz</i>!</h1>
+                  <p>
+                    <i>archeoViz</i> ist eine für die Archäologie entwickelte Anwendung. 
+                    Sie ermöglicht die <b>Visualisierung</b>, interaktive <b>Erkundung</b> und 
+                    schnelle Bereitstellung und <b>Kommunikation</b> von Raumdaten
+                    archäologischer Grabungen und Prospektionen im Internet.
+                  </p>
+                  <p>
+                    <i>archeoViz</i> bietet interaktive 3D und 2D <b>Visualisierungen</b> und kann
+                    <b>Profile</b>, als auch <b>Karten</b> der Fundverteilungen generieren. Es 
+                    bietet zudem die Möglichkeit grundlegende Methoden der <b>räumlichen Statistik</b>
+                    ausführen (konvexe Hülle, Regressionsoberflächen, 2D-Kernel-Dichte-Schätzung)
+                    und stellt eine interaktive Darstellung der Ausgrabung in Zeitschritten zur 
+                    Verfügung. 
+                  </p> 
+                  <p>
+                    Als R-Paket, kann <i>archeoViz</i> lokal mit eigenen Daten ausgeführt werden.
+                    Oder Daten können Online über eine Eingabemaske eingeladen werden. Die 
+                    Eingabemaske ist derezit in Englisch, Französisch, Italienisch, Portugiesisch und 
[…]
