+++ R/load_interface_terms_utf8.R
@@ -5,7 +5,7 @@
   welcome.fr <- "<h1>Bienvenue dans <i>archeoViz</i> !</h1>
                 <p><i>archeoViz</i> est une application dédiée à l'archéologie.
                 Elle permet de <b>visualiser</b>, d'<b>explorer</b> interactivement, et  
-                d'exposer et <b>communiquer</b> rapidement  sur le web des données 
+                d'exposer et <b>communiquer</b> rapidement sur le web des données 
                 archéologiques de terrain.
                 </p>
                 <p>
@@ -28,7 +28,7 @@
                 <h1>Welcome to <i>archeoViz</i>!</h1>
                 <p><i>archeoViz</i> is an application designed for archaeology.
                 It makes possible to <b>visualise</b>, interactively <b>explore</b>, and quickly
-                deploy and <b>communicate</b> archaeological excavation data on the web.
+                deploy and <b>communicate</b> archaeological spatial data on the web.
                 </p>
                 <p>
                 It offers interactive 3D and 2D <b>visualisations</b>,
@@ -38,7 +38,7 @@
                 and display an interactive <b>timeline</b> of the excavation. 
                 </p>
                 <p>
-                Being wrapped as an R package, 
+                Shared as an R package, 
                 <i>archeoViz</i> can be used locally or deployed on a server, 
                 either by allowing the user to load data through the interface
                 or by running the app with a specific data set.
@@ -80,7 +80,7 @@
       <li> <b>id</b> : <i>valeur alphanumérique</i>, identifiant unique de l'objet </li>
       <li> <b>xmin</b> : <i>valeur numérique</i>, coordonnée de l'objet en axe X (en cm)</li>
       <li> <b>ymin</b> : <i>valeur numérique</i>, coordonnée de l'objet en axe Y (en cm)</li>
-      <li> <b>zmin</b> : <i>valeur numérique</i>, coordonnée de l'objet en axe Z (valeur positive de profondeur en cm)</li>
+      <li> <b>zmin</b> : <i>valeur numérique</i>, coordonnée de l'objet en axe Z (valeur de profondeur en cm)</li>
       <li> <b>layer</b> : <i>valeur alphanumérique</i>, identifiant de la couche de l'objet</li>
       <li> <b>object_type</b> : <i>valeur alphanumérique</i>, catégorie de l'objet</li>
     </ul>
@@ -104,7 +104,7 @@
       <li> <b>id</b>: <i>alphanumerical value</i>, unique identifier of the object </li>
       <li> <b>xmin</b>: <i>numerical value</i>, coordinate of the object on the X axis (in cm)</li>
       <li> <b>ymin</b>: <i>numerical value</i>, coordinate of the object on the Y axis (in cm)</li>
-      <li> <b>zmin</b>: <i>numerical value</i>, coordinate of the object on the Z axis (positive depth value in cm)</li>
+      <li> <b>zmin</b>: <i>numerical value</i>, coordinate of the object on the Z axis (depth value in cm)</li>
       <li> <b>layer</b>: <i>alphanumerical value</i>, identifier of the object's layer</li>
       <li> <b>object_type</b>: <i>alphanumerical value</i>, category of the object</li>
     </ul>
@@ -128,7 +128,7 @@
       <li> <b>id</b> : <i>valor alfanumérico</i>, identificador único do objeto </li>
       <li> <b>xmin</b> : <i>valor numérico</i>, coordenada do objeto no eixo X (em cm)</li>
       <li> <b>ymin</b> : <i>valor numérico</i>, coordenada do objeto no eixo Y (em cm)</li>
[…]
