+++ paper.bib
@@ -73,12 +73,12 @@
 }
 
 
-@Manual{Plutniak2023archeoviz1.1.1,
+@Manual{Plutniak2023archeoviz1.2.1,
   Title                    = {archeoViz. Visualisation, Exploration, and Web Communication of Archaeological Spatial Data},
-  Author                   = {Plutniak, Sébastien and Araujo, Renata and Giardino, Sara},
+  Author                   = {Plutniak, Sébastien and Araujo, Renata and Giardino, Sara and Laabs, Julian},
   Year                     = {2023},
-  Doi                      = {10.5281/zenodo.7938998},
-  Note                     = {R package version 1.1.1},
+  Doi                      = {10.5281/zenodo.7460193},
+  Note                     = {R package version 1.2.1},
   Url                      = {https://CRAN.R-project.org/package=archeoViz},
   Langid                   = {american}
 }
