+++ R/load_interface_terms_utf8.R
@@ -20,7 +20,7 @@
                 <i>archeoViz</i> peut être utilisée localement ou déployée sur un serveur,
                 soit en chargeant des données via l'interface, soit en 
                 lançant l'application avec un jeu de donnée spécifique.
-                L'interface est disponible en anglais, français, italien et portugais.
+                L'interface est disponible en <a href=https://analytics.huma-num.fr/archeoviz/de target=_blank>allemand</a>, <a href=https://analytics.huma-num.fr/archeoviz/en target=_blank>anglais</a>, français, <a href=https://analytics.huma-num.fr/archeoviz/it target=_blank>italien</a> et <a href=https://analytics.huma-num.fr/archeoviz/pt target=_blank>portugais</a>.
                 </p>
                 Le package, sous licence libre, est publié sur le <a href=https://cran.r-project.org/package=archeoViz target=_blank>CRAN</a>. Son code source est aussi disponible sur <a href=https://github.com/sebastien-plutniak/archeoviz target=_blank>github</a>."
   
@@ -43,7 +43,7 @@
                 <i>archeoViz</i> can be used locally or deployed on a server, 
                 either by allowing the user to load data through the interface
                 or by running the app with a specific data set.
-                The interface is available in English, French, Italian and Portuguese.
+                The interface is available in <a href=https://analytics.huma-num.fr/archeoviz/de target=_blank>German</a>, English, <a href=https://analytics.huma-num.fr/archeoviz/fr target=_blank>French</a>, <a href=https://analytics.huma-num.fr/archeoviz/it target=_blank>Italian</a>, and <a href=https://analytics.huma-num.fr/archeoviz/pt target=_blank>Portuguese</a>.
                 </p>
                 The package is published under a free license on the <a href=https://cran.r-project.org/package=archeoViz target=_blank>CRAN</a>. Its source code is also available on  <a href=https://github.com/sebastien-plutniak/archeoviz target=_blank>github</a>."
   
@@ -57,7 +57,8 @@
                 </p>
                 <p>
                 Distribuído como um pacote R, 
-                o <i>archeoViz</i> pode ser usado localmente ou implantado em um servidor, seja carregando dados por meio da interface, seja iniciando o aplicativo com um conjunto de dados específico. A interface está disponível em inglês, francês, italiano, e português.
+                o <i>archeoViz</i> pode ser usado localmente ou implantado em um servidor, seja carregando dados por meio da interface, seja iniciando o aplicativo com um conjunto de dados específico. A interface está disponível em <a href=https://analytics.huma-num.fr/archeoviz/de target=_blank>alemão</a>
+<a href=https://analytics.huma-num.fr/archeoviz/en target=_blank>inglês</a>, <a href=https://analytics.huma-num.fr/archeoviz/fr target=_blank>francês</a>, <a href=https://analytics.huma-num.fr/archeoviz/it target=_blank>italiano</a>, e português.
                 </p>
[…]
