+++ R/load_interface_terms.R
@@ -3,32 +3,33 @@
   # welcome ----
   # : fr ----
   welcome.fr <- "<h1>Bienvenue dans <i>archeoViz</i> !</h1>
-                <p><i>archeoViz</i> est une application d\\u00e9di\\u00e9e \\u00e0 l'arch\\u00e9ologie.
+                <p><i>archeoViz</i> est une application d\\\\u00e9di\\\\u00e9e \\\\u00e0 l'arch\\\\u00e9ologie.
                 Elle permet de <b>visualiser</b>, d'<b>explorer</b> interactivement, et  
-                d'exposer et <b>communiquer</b> rapidement  sur le web des donn\\u00e9es 
-                arch\\u00e9ologiques de terrain.
+                d'exposer et <b>communiquer</b> rapidement sur le web des donn\\\\u00e9es 
+                arch\\\\u00e9ologiques de terrain.
                 </p>
                 <p>
                 Elle propose des <b>visualisations</b> en 3D et 2D,
-                g\\u00e9n\\u00e8re des <b>coupes</b> et des <b>cartes</b> des restes arch\\u00e9ologiques,
-                permet de r\\u00e9aliser des <b>statistiques spatiales</b> simples
-                (enveloppes convexes, surfaces de r\\u00e9gression, estimation de densit\\u00e9 par noyau en 2D), 
+                g\\\\u00e9n\\\\u00e8re des <b>coupes</b> et des <b>cartes</b> des restes arch\\\\u00e9ologiques,
+                permet de r\\\\u00e9aliser des <b>statistiques spatiales</b> simples
+                (enveloppes convexes, surfaces de r\\\\u00e9gression, estimation de densit\\\\u00e9 par noyau en 2D), 
                 et de visualiser une <b>chronologie</b> interactive des fouilles d'un site.
                 </p>
                 <p>
-                Distribu\\u00e9e sous la forme d'un package R, 
-                <i>archeoViz</i> peut \\u00eatre utilis\\u00e9e localement ou d\\u00e9ploy\\u00e9e sur un serveur,
-                soit en chargeant des donn\\u00e9es via l'interface, soit en 
-                lan\\u00e7ant l'application avec un jeu de donn\\u00e9e sp\\u00e9cifique.
-                L'interface est disponible en anglais, fran\\u00e7ais, italien et portugais.
+                Distribu\\\\u00e9e sous la forme d'un package R, 
+                <i>archeoViz</i> peut \\\\u00eatre utilis\\\\u00e9e localement ou d\\\\u00e9ploy\\\\u00e9e sur un serveur,
+                soit en chargeant des donn\\\\u00e9es via l'interface, soit en 
+                lan\\\\u00e7ant l'application avec un jeu de donn\\\\u00e9e sp\\\\u00e9cifique.
+                L'interface est disponible en <a href=https://analytics.huma-num.fr/archeoviz/de target=_blank>allemand</a>, <a href=https://analytics.huma-num.fr/archeoviz/en target=_blank>anglais</a>, fran\\\\u00e7ais, <a href=https://analytics.huma-num.fr/archeoviz/it target=_blank>italien</a> et <a href=https://analytics.huma-num.fr/archeoviz/pt target=_blank>portugais</a>.
                 </p>
-                Le package, sous licence libre, est publi\\u00e9 sur le <a href=https://cran.r-project.org/package=archeoViz target=_blank>CRAN</a>. Son code source est aussi disponible sur <a href=https://github.com/sebastien-plutniak/archeoviz target=_blank>github</a>."
[…]
