+++ vignettes/archeoViz.Rmd
@@ -130,6 +130,7 @@ The language of the application can be set with the `lang` parameter.
 Demonstration instances of the application are deployed on the *Huma Num* Shiny server:
 
 * [`archeoViz` in English](https://analytics.huma-num.fr/archeoviz/en).
+* [`archeoViz` in German](https://analytics.huma-num.fr/archeoviz/de).
 * [`archeoViz` in French](https://analytics.huma-num.fr/archeoviz/fr).
 * [`archeoViz` in Italian](https://analytics.huma-num.fr/archeoviz/it).
 * [`archeoViz` in Portuguese](https://analytics.huma-num.fr/archeoviz/pt).
@@ -536,5 +537,5 @@ The `archeoViz` application and package is developed and maintained by Sébastie
 
 # References
 
-* Plutniak, Sébastien, Renata Araujo, Sara Giardino. 2023. “archeoViz. Visualisation, Exploration, and Web Communication of Archaeological Spatial Data”. v1.2.0, DOI: [10.5281/zenodo.7460193](https://doi.org/10.5281/zenodo.7460193).
+* Plutniak, Sébastien, Renata Araujo, Sara Giardino, Julian Laabs. 2023. “archeoViz. Visualisation, Exploration, and Web Communication of Archaeological Spatial Data”. v1.2.1, DOI: [10.5281/zenodo.7460193](https://doi.org/10.5281/zenodo.7460193).
 * Plutniak, Sébastien. 2023. “[Visualiser et explorer la distribution spatiale du mobilier archéologique : l'application archeoViz et son portail web](https://www.prehistoire.org/offres/doc_inline_src/515/0-BSPF_2023_1_2e_partie_Correspondance_PLUTNIAK.pdf)”. *Bulletin de la Société préhistorique française*, 120(1), p. 70-74.

[…]
