+++ paper.bib
@@ -77,7 +77,7 @@
   Title                    = {archeoViz. Visualisation, Exploration, and Web Communication of Archaeological Spatial Data},
   Author                   = {Plutniak, Sébastien and Araujo, Renata and Giardino, Sara and Laabs, Julian},
   Year                     = {2023},
-  Doi                      = {10.5281/zenodo.7460193},
+  Doi                      = {10.5281/zenodo.8092474},
   Note                     = {R package version 1.2.1},
   Url                      = {https://CRAN.R-project.org/package=archeoViz},
   Langid                   = {american}
