+++ vignettes/archeoViz.Rmd
@@ -15,7 +15,7 @@ knitr::opts_chunk$set(
 ```
 
 
-`archeoViz` is a packaged R Shiny application for the *visualisation*, *exploration*, and web *communication* of archaeological spatial data. It includes interactive 3D and 2D *visualisations*, can generate *cross sections* and *maps* of the remains, can run basic *spatial statistics* methods (convex hull, regression surfaces, 2D kernel density estimation), and display an interactive *timeline* of an excavation. `archeoViz` can be used locally or deployed on a server, either by allowing the user to load data through the interface or by running the app with a specific data set. The app interface is available in English, French, Italian, and Portuguese.
+`archeoViz` is a packaged R Shiny application for the *visualisation*, *exploration*, and web *communication* of archaeological spatial data. It includes interactive 3D and 2D *visualisations*, can generate *cross sections* and *maps* of the remains, can run basic *spatial statistics* methods (convex hull, regression surfaces, 2D kernel density estimation), and display an interactive *timeline* of an excavation. `archeoViz` can be used locally or deployed on a server, either by allowing the user to load data through the interface or by running the app with a specific data set. The app interface is available in English, French, German, Italian, and Portuguese.
 
 
 
