+++ vignettes/archeoViz.Rmd
@@ -18,7 +18,6 @@ knitr::opts_chunk$set(
 `archeoViz` is a packaged R Shiny application for the *visualisation*, *exploration*, and web *communication* of archaeological spatial data. It includes interactive 3D and 2D *visualisations*, can generate *cross sections* and *maps* of the remains, can run basic *spatial statistics* methods (convex hull, regression surfaces, 2D kernel density estimation), and display an interactive *timeline* of an excavation. `archeoViz` can be used locally or deployed on a server, either by allowing the user to load data through the interface or by running the app with a specific data set. The app interface is available in English, French, German, Italian, and Portuguese.
 
 
-
 - [**Installation**](#installation)
   - [Local Use](#local-use)
   - [Remote Use](#remote-use)
@@ -29,7 +28,7 @@ knitr::opts_chunk$set(
 - [**Use**](#use)
   - [Spatial Information in archeoViz](#spatial-information-in-archeoviz)
     - [Exact Location: Plotted Objects](#exact-location-plotted-objects)
-    - [Vague Location: Spits, Buckets, Sieved Objects, and Errors](#vague-location-spits-buckets-sieved-objects-and-errors)
+    - [Vague Location: Spits, Buckets, Sieved Objects, and Recording Errors](#vague-location-spits-buckets-sieved-objects-and-recording-errors)
   - [Refittings and Fabric Measurements](#refittings-and-fabric-measurements)
       - [Refittings](#refittings)
       - [Fabric Measurements](#fabric-measurements)      
@@ -129,8 +128,8 @@ The language of the application can be set with the `lang` parameter.
 
 Demonstration instances of the application are deployed on the *Huma Num* Shiny server:
 
-* [`archeoViz` in English](https://analytics.huma-num.fr/archeoviz/en).
 * [`archeoViz` in German](https://analytics.huma-num.fr/archeoviz/de).
+* [`archeoViz` in English](https://analytics.huma-num.fr/archeoviz/en).
 * [`archeoViz` in French](https://analytics.huma-num.fr/archeoviz/fr).
 * [`archeoViz` in Italian](https://analytics.huma-num.fr/archeoviz/it).
 * [`archeoViz` in Portuguese](https://analytics.huma-num.fr/archeoviz/pt).
@@ -171,7 +170,7 @@ Archaeologists record the location of archaeological objects at different scales
 Using grid coordinates or electronic “total station” enables recording  the individual location of objects on the field.
 In that case, the location are represented as points in `archeoViz` (triplets of x, y and z coordinates).
 
-### Vague Location: Spits, Buckets, Sieved Objects, and Errors
+### Vague Location: Spits, Buckets, Sieved Objects, and Recording Errors
 
 However, it is also common that x, y, z, coordinates by object are not available,  for different reasons due to:
 
@@ -217,7 +216,7 @@ Three types of data can be loaded in `archeoViz`:
 * a “refits” table (optional), with data about the refitting objects;
 * a “timeline” table (optional), with data about when each square of the site was excavated or surveyed.
 
-### Formatting data
+### Formatting Data
 
 The tables must be CSV files with the first row containing the column labels. 
 Contents in HTML are allowed. This makes it possible, in particular, to add links to external resources (e.g., to objects permanent identifier in other databases, or to concepts identifiers in standard ontologies / thesauri, etc.).
@@ -227,7 +226,7 @@ Formatting your data can be done:
 * either using a spreadsheet editor on your machine to generate CSV files;
[…]
