+++ README-fr.md
@@ -795,7 +795,7 @@ l’application respectivement en portugais, italien et allemand.
     [10.5281/zenodo.7460193](https://doi.org/10.5281/zenodo.7460193).
   - Plutniak, Sébastien, Anaïs Vignoles. 2023. “[The archeoViz Portal:
     Dissemination of Spatial Archaeological
-    Datasets](https://hal.science/hal-04156271), portail web,
+    Datasets](https://hal.science/hal-04156271)”, portail web,
     <https://analytics.huma-num.fr/archeoviz/home>.
 
 ## Articles
