+++ vignettes/archeoViz.Rmd
@@ -539,7 +539,7 @@ The `archeoViz` application and package is developed and maintained by Sébastie
 ## Software
 
 * Plutniak, Sébastien, Renata Araujo, Sara Giardino, Julian Laabs. 2023. “archeoViz. Visualisation, Exploration, and Web Communication of Archaeological Spatial Data”. v1.2.1, DOI: [10.5281/zenodo.7460193](https://doi.org/10.5281/zenodo.7460193).
-* Plutniak, Sébastien, Anaïs Vignoles. 2023. “[The archeoViz Portal: Dissemination of Spatial Archaeological Datasets](https://hal.science/hal-04156271), portail web, https://analytics.huma-num.fr/archeoviz/home.
+* Plutniak, Sébastien, Anaïs Vignoles. 2023. “[The archeoViz Portal: Dissemination of Spatial Archaeological Datasets](https://hal.science/hal-04156271)”, portail web, https://analytics.huma-num.fr/archeoviz/home.
 
 ## Papers
 
