### 2023-07-12 #5578 @martinfleis
I am overbooked right now but wanted to link to a last year's related submission by @sebastien-plutniak https://github.com/openjournals/joss-reviews/issues/4335 as it may help a future editor.
