+++ R/load_interface_terms.R
@@ -814,7 +814,7 @@ guidelines.fr <- "
 </ul></li>
 </ul></li>
 <li><a href=#reproductibilit\\\\u00e9><strong>Reproductibilit\\\\u00e9</strong></a></li>
-<li><a href=#exports-depuis-et-vers-des-applications-tiers><strong>Exports depuis et vers des applications tiers</strong></a>
+<li><a href=#exports-depuis-et-vers-des-applications-tierces><strong>Exports depuis et vers des applications tierces</strong></a>
 <ul>
 <li><a href=#export-depuis-archeoviz>Export depuis archeoViz</a></li>
 <li><a href=#import-vers-archeoviz>Import vers archeoViz</a></li>
@@ -1041,7 +1041,7 @@ guidelines.fr <- "
 <li>Dans l\\\\u2019onglet \\\\u201cReproductibilit\\\\u00e9\\\\u201d, une commande R est g\\\\u00e9n\\\\u00e9r\\\\u00e9e dynamiquement, tenant compte du param\\\\u00e9trage de l\\\\u2019application r\\\\u00e9alis\\\\u00e9 par l\\\\u2019utilisateur en agissant avec l\\\\u2019interface graphique.</li>
 <li>Dans un usage plus avanc\\\\u00e9, les param\\\\u00e8tres d\\\\u2019URL permettent de param\\\\u00e9trer une instance en ligne de l\\\\u2019application avec des param\\\\u00e8tres d\\\\u2019int\\\\u00e9r\\\\u00eat et de la communiquer l\\\\u2019ensemble en envoyant l\\\\u2019URL \\\\u00e0 un tiers.</li>
 </ul>
-<h2 id=exports-depuis-et-vers-des-applications-tiers>Exports depuis et vers des applications tiers</h2>
+<h2 id=exports-depuis-et-vers-des-applications-tierces>Exports depuis et vers des applications tierces</h2>
 <p><code>archeoViz</code> a \\\\u00e9t\\\\u00e9 con\\\\u00e7u comme l\\\\u2019une des pi\\\\u00e8ces d\\\\u2019un \\\\u00e9cosyst\\\\u00e8me num\\\\u00e9rique d\\\\u00e9centralis\\\\u00e9 pour les donn\\\\u00e9es et analyses arch\\\\u00e9ologiques. Dans cette approche, les fonctionnalit\\\\u00e9s sont distribu\\\\u00e9es entre de multiples applications interconenct\\\\u00e9es, plut\\\\u00f4t que concentr\\\\u00e9es dans un petit nombre de syst\\\\u00e8mes. Par cons\\\\u00e9quent, les donn\\\\u00e9es peuvent \\\\u00eatre export\\\\u00e9es et import\\\\u00e9es entre <code>archeoViz</code> et d\\\\u2019autres applications web. Notez que, jusqu\\\\u2019ici, les fonctionnalit\\\\u00e9s d\\\\u2019export ne sont disponibles que dans le cas d\\\\u2019instances <code>archeoViz</code> d\\\\u00e9ploy\\\\u00e9es en ligne.</p>
 <h3 id=export-depuis-archeoviz>Export depuis archeoViz</h3>
[…]
