### 2023-07-25 #5578 @Bisaloo
Hi, thanks for the invite. I'll be on annual leave over the next month and prefer not taking any assignment that will inevitably be delayed as a result. I'll update my availability on the editor dashboard to reflect this.
