+++ R/load_interface_terms_utf8.R
@@ -1187,6 +1187,8 @@ guidelines.fr <- "
     tab.input = "Données",
     tab.plot3d ="Vue 3D",
     tab.map = "Plan",
+    tab.section.x = "Section X",
+    tab.section.y = "Section Y",
     tab.statistics = "Statistiques",
     tab.timeline = "Chronologie",
     tab.reproducibility = "Reproductibilité",
@@ -1262,6 +1264,8 @@ guidelines.fr <- "
     tab.input = "Input data",
     tab.plot3d ="3D plot",
     tab.map = "Map",
+    tab.section.x = "Section X",
+    tab.section.y = "Section Y",
     tab.statistics = "Statistics",
     tab.timeline = "Timeline",
     tab.reproducibility = "Reproducibility",
@@ -1336,6 +1340,8 @@ guidelines.fr <- "
     tab.input = "Dados",
     tab.plot3d ="Visualização 3D",
     tab.map = "Mapa",
+    tab.section.x = "Perfil X",
+    tab.section.y = "Perfil Y",
     tab.statistics = "Estatísticas",
     tab.timeline = "Cronologia",
     tab.reproducibility = "Reprodutibilidade",
@@ -1411,6 +1417,8 @@ guidelines.fr <- "
     tab.input = "Dati",
     tab.plot3d = "Vista 3D",
     tab.map = "Mappa",
+    tab.section.x = "Sezione X",
+    tab.section.y = "Sezione Y",
     tab.statistics = "Statistiche",
     tab.timeline = "Cronologia",
     tab.reproducibility = "Riproducibilità",
@@ -1486,6 +1494,8 @@ guidelines.fr <- "
     tab.input = "Dateneingabe",
     tab.plot3d ="3D Ansicht",
     tab.map = "Karte",
+    tab.section.x = "Profile X",
+    tab.section.y = "Profile Y",
     tab.statistics = "Statistiken",
     tab.timeline = "Chronologie",
     tab.reproducibility = "Reproduzierbarkeit",
