+++ R/archeoViz.R
@@ -73,7 +73,7 @@ archeoViz <- function(objects.df = NULL, refits.df = NULL, timeline.df = NULL,
                "params" = params,
                "title"       = title,
                "home.text"   = home.text,
-               "lang"        = lang,
+               # "lang"        = lang,
                "set.theme"   = set.theme,
                "ui.terms"    = ui.terms,
                "background.col" = background.col,
