### 2023-08-02 d4318e4
stat tables must depend on button and run.plot
