### 2023-08-03 78b9d4d
fix run.plots behavior, reproducibility bug, add point.size param, color lines
