### 2023-08-18 #5578 @oliviaguest
Hi @joeroe, @bischrob, @nevrome, do you have the time to review this for JOSS? :sparkles:
