### 2023-08-18 #5578 @bischrob
Hi @oliviaguest , I would be happy to review this. What is the process for JOSS?
