### 2023-08-18 #5578 @oliviaguest
@bischrob please check this and let me know if you have any questions, but bear in mind the review has not started yet! https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
