### 2023-08-21 #5578 @joeroe
I reviewed Sebastien's previous JOSS paper last year (#4335), but if that's not a problem I could also review this one.
