### 2023-08-21 #5578 @bischrob
> @bischrob please check this and let me know if you have any questions, but bear in mind the review has not started yet! https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

@oliviaguest, this is very helpful, thanks!
