+++ R/load_interface_terms.R
@@ -102,7 +102,29 @@
                     <a href=https://github.com/sebastien-plutniak/archeoviz target=_blank>github</a> 
                       verf\\\\u00fcgbar.
                   </p>"
-
+  # : es ----
+  welcome.es <- "<h1>Bienvenidos en <i>archeoViz</i> !</h1>
+                <p><i>archeoViz</i> es una aplicaci\\\\u00f3n dedicada al arqueolog\\\\u00eda.
+                Permite <b>visualizar</b>, <b>explorar</b> de manera interactiva, y 
+                presentar y <b>comunicar</b> datos arqueol\\\\u00f3gicos de campo r\\\\u00e1pidamente en la web.
+                </p>
+                <p>
+                Ofrece <b>visualizaciones</b> en 3D y 2D,
+                genera <b>perfiles estratigr\\\\u00e1ficos</b> y <b>mapas</b> de los vestigios arqueol\\\\u00f3gicos,
+                permite correr <b>estad\\\\u00edsticas espaciales</b> sencillas
+                (cascos convexos, superficies de regresi\\\\u00f3n, estimaci\\\\u00f3n de densidad por n\\\\u00facleo en 2D), 
+                y visualizar una <b>cronolog\\\\u00eda</b> interactiva de las excavaciones de un sitio.
+                </p>
+                <p>
+                Distribuida como un paquete R, 
+                <i>archeoViz</i> se puede usar localmente o implementada en un servidor,
+                sea cargando los datos trav\\\\u00e9s de la interfaz, o
+                 iniciando la aplicaci\\\\u00f3n con un conjunto de datos espec\\\\u00edfico.
+                La interfaz est\\\\u00e1 disponible en <a href=https://analytics.huma-num.fr/archeoviz/de target=_blank>alem\\\\u00e1n</a>, <a href=https://analytics.huma-num.fr/archeoviz/en target=_blank>ingl\\\\u00e9s</a>, franc\\\\u00e9s, <a href=https://analytics.huma-num.fr/archeoviz/it target=_blank>italiano</a>, <a href=https://analytics.huma-num.fr/archeoviz/pt target=_blank>portugu\\\\u00e9s</a>, y <a href=https://analytics.huma-num.fr/archeoviz/es target=_blank>espa\\\\u00f1ol</a>.
+                </p>
+                El paquete, bajo licencia libre, est\\\\u00e1 publicado en <a href=https://cran.r-project.org/package=archeoViz target=_blank>CRAN</a>. Su c\\\\u00f3digo fuente tambi\\\\u00e9n est\\\\u00e1 disponible en <a href=https://github.com/sebastien-plutniak/archeoviz target=_blank>github</a>."
+  
+  
   
   # Input objects ----
   # :  fr ----
@@ -225,6 +247,29 @@
       </ul>
       </p>"
 
+# :  es ----
+  input.objects.es <- 
+    "<p>
+    Se necesita una tabla bajo el formato csv. Cada l\\\\u00ednea describe un objeto, incluyendo los siguientes campos obligatorios:
+    <ul>
+      <li> <b>id</b> : <i>valor alfanum\\\\u00e9rico</i>, identificador \\\\u00fanico del objeto </li>
+      <li> <b>xmin</b> : <i>valor num\\\\u00e9rico</i>, coordenada del objeto en el eje X (en cm)</li>
+      <li> <b>ymin</b> : <i>valor num\\\\u00e9rico</i>, coordenada del objeto en el eje Y (en cm)</li>
+      <li> <b>zmin</b> : <i>valor num\\\\u00e9rico</i>, coordenada del objeto en el eje Z (valor de profundidad en cm)</li>
[…]
