+++ R/load_interface_terms_utf8.R
@@ -102,7 +102,29 @@
                     <a href=https://github.com/sebastien-plutniak/archeoviz target=_blank>github</a> 
                       verfügbar.
                   </p>"
-
+  # : es ----
+  welcome.es <- "<h1>Bienvenidos en <i>archeoViz</i> !</h1>
+                <p><i>archeoViz</i> es una aplicación dedicada al arqueología.
+                Permite <b>visualizar</b>, <b>explorar</b> de manera interactiva, y 
+                presentar y <b>comunicar</b> datos arqueológicos de campo rápidamente en la web.
+                </p>
+                <p>
+                Ofrece <b>visualizaciones</b> en 3D y 2D,
+                genera <b>perfiles estratigráficos</b> y <b>mapas</b> de los vestigios arqueológicos,
+                permite correr <b>estadísticas espaciales</b> sencillas
+                (cascos convexos, superficies de regresión, estimación de densidad por núcleo en 2D), 
+                y visualizar una <b>cronología</b> interactiva de las excavaciones de un sitio.
+                </p>
+                <p>
+                Distribuida como un paquete R, 
+                <i>archeoViz</i> se puede usar localmente o implementada en un servidor,
+                sea cargando los datos través de la interfaz, o
+                 iniciando la aplicación con un conjunto de datos específico.
+                La interfaz está disponible en <a href=https://analytics.huma-num.fr/archeoviz/de target=_blank>alemán</a>, <a href=https://analytics.huma-num.fr/archeoviz/en target=_blank>inglés</a>, francés, <a href=https://analytics.huma-num.fr/archeoviz/it target=_blank>italiano</a>, <a href=https://analytics.huma-num.fr/archeoviz/pt target=_blank>portugués</a>, y <a href=https://analytics.huma-num.fr/archeoviz/es target=_blank>español</a>.
+                </p>
+                El paquete, bajo licencia libre, está publicado en <a href=https://cran.r-project.org/package=archeoViz target=_blank>CRAN</a>. Su código fuente también está disponible en <a href=https://github.com/sebastien-plutniak/archeoviz target=_blank>github</a>."
+  
+  
   
   # Input objects ----
   # :  fr ----
@@ -225,6 +247,29 @@
       </ul>
       </p>"
 
+# :  es ----
+  input.objects.es <- 
+    "<p>
+    Se necesita una tabla bajo el formato csv. Cada línea describe un objeto, incluyendo los siguientes campos obligatorios:
+    <ul>
+      <li> <b>id</b> : <i>valor alfanumérico</i>, identificador único del objeto </li>
+      <li> <b>xmin</b> : <i>valor numérico</i>, coordenada del objeto en el eje X (en cm)</li>
+      <li> <b>ymin</b> : <i>valor numérico</i>, coordenada del objeto en el eje Y (en cm)</li>
+      <li> <b>zmin</b> : <i>valor numérico</i>, coordenada del objeto en el eje Z (valor de profundidad en cm)</li>
+      <li> <b>layer</b> : <i>valor alfanumérico</i>, identificador de capa del objeto</li>
+      <li> <b>object_type</b> : <i>valor alfanumérico</i>, categoría del objeto</li>
+    </ul>
+    Además, otros campos opcionales son posibles, incluyendo:
+    <ul>
+      <li> <b>square_x</b> : <i>valor alfanumérico</i>, identificador del cuadrado del objeto en el eje X</li>
+      <li> <b>square_y</b> : <i>valor alfanumérico</i>, identificador del cuadrado del objeto en el eje Y</li>
+      <li> <b>year</b> : <i>valor numérico</i>, año de excavación del objeto</li>
[…]
