### 2023-09-07 #5811 @oliviaguest
:wave: **Hi @bischrob, @nevrome, @joeroe**, thank you so much for helping out at JOSS. If you need any pointers, please feel free to look at previous reviews (which can be found by looking at published papers) and the [documentation](https://joss.readthedocs.io/). If you need to comment on the code itself, opening an issue at the repo and then linking to it from here (to help me/others keep track) is the way to go. For comments on the paper, you can also open issues or PRs (say for typos), but those can be directly posted as replies in this issue. Thanks, and feel free to reach out if you need me. :relaxed:
