### 2023-09-13 #5811 @bischrob
@oliviaguest -- the submitting author was a visiting scholar at my institution (Arizona State University) last summer. We met for lunch to discuss our similar interests but worked in different organizations and have not collaborated. I do not consider this a conflict of interest. Can you confirm?
