+++ R/app_server.R
@@ -1,6 +1,6 @@
 app_server <- function(input, output, session) {
   .data <- NULL
-
+  
   # retrieve parameters from URL ----
   observe({
     query <- shiny::parseQueryString(session$clientData$url_search)
@@ -48,7 +48,7 @@ app_server <- function(input, output, session) {
           )))
         }
       }
-
+      
       # coerce to logical values:
       sel <- param.list %in% c("TRUE", "FALSE", "T", "F")
       param.list[ sel ] <- as.logical(param.list[ sel ])
@@ -56,7 +56,7 @@ app_server <- function(input, output, session) {
       shinyOptions("params" =  param.list)
     }
   }, priority=10)
-    
+  
   
   
   # Interface ----
@@ -225,7 +225,7 @@ app_server <- function(input, output, session) {
     showNotification(.term_switcher(result$notif.text),
                      type = result$notif.type, duration = 10)
     if(result$notif.type == "error") return(NULL)
-
+    
     result$data
   })
   
@@ -649,7 +649,7 @@ app_server <- function(input, output, session) {
     
     # : add refits lines  ----
     plot3d.refits <- sum(c(input$plot3d.refits,
-                          getShinyOption("params")$plot3d.refits))
+                           getShinyOption("params")$plot3d.refits))
     if( plot3d.refits > 0 ){
       refitting.df <- refitting.df()
       refitting.df <- refitting.df$refits.3d
@@ -795,7 +795,7 @@ app_server <- function(input, output, session) {
     } # end if("show.uncertainty" %in% input$location)
     
     
-    # : add surfaces ####
+    # : add surfaces ----
     if(sum(input$plot3d.surfaces) > 0){
       
       # filter the layers for which a regression surfaces must be computed:
@@ -822,14 +822,18 @@ app_server <- function(input, output, session) {
     
     # :  add convex hull ####
     if(input$plot3d.hulls){
-      # filter the layers for which a convex hull must be computed:
-      subsets <- table(dataset$group.variable) 
-      subsets <- names(subsets[subsets > 19])
       
       # compute hulls:
-      mesh.list <- lapply(subsets, function(x) 
-        .get_cxhull_model(df=dataset,
-                          var=group.variable(), 
+      
+      hulls.selection <- input$hulls.selection
+      # for first launch (with run.plot=T) :
+      if(is.null(hulls.selection)){ 
+        hulls.selection <- hulls.list()$hull.selected
+      }
+      
+      mesh.list <- lapply(hulls.selection, function(x) 
+        .get_cxhull_model(df = dataset,
+                          var = group.variable(), 
                           value = x))
       # add convex hull mesh:
       for(i in seq_len(length(mesh.list)) ){
@@ -868,7 +872,7 @@ app_server <- function(input, output, session) {
     
     # camera settings:
     camera.values <- list()
-
+    
     center.xyz <- getShinyOption("params")$camera.center
     camera.values$center <- list(x = center.xyz[1],
                                  y = center.xyz[2], z = center.xyz[3])
@@ -991,7 +995,7 @@ app_server <- function(input, output, session) {
     } else{
       sectionY.point.size <- input$sectionY.point.size
     }
-
+    
     # sel <- (dataset$y >= input$sectionY.y.val[1] & dataset$y <= input$sectionY.y.val[2]) &
     #   (dataset$x >= input$sectionY.x.val[1] & dataset$x <= input$sectionY.x.val[2])
     
@@ -1023,7 +1027,7 @@ app_server <- function(input, output, session) {
       htmlwidgets::saveWidget(sectionYplot(), file = file3)
     }
   )
- # : Map plot ----
+  # : Map plot ----
   map <- eventReactive(input$goButtonZ, {
[…]
