+++ .github/workflows/r.yml
@@ -47,7 +47,7 @@ jobs:
         if: matrix.config.os == 'macOS-latest'
       - name: Install R dependencies (macOs)
         run: |
-          install.packages(c("remotes", "rcmdcheck", "plotly"))
+          install.packages(c("remotes", "rcmdcheck", "plotly", "SEAHORS"))
           remotes::install_deps(dependencies = TRUE)
         if: matrix.config.os == 'macOS-latest'          
         shell: Rscript {0}    
@@ -73,7 +73,7 @@ jobs:
         shell: Rscript {0}              
       - name: Install R dependencies (Ubuntu)
         run: |
-          install.packages(c("remotes", "curl", "pkgbuild", "rcmdcheck", "purrr", "sf", "Morpho", "cxhull", "Rvcg"))
+          install.packages(c("remotes", "curl", "pkgbuild", "rcmdcheck", "purrr", "sf", "Morpho", "cxhull", "Rvcg", "SEAHORS"))
           remotes::install_deps(dependencies = TRUE)
         if: matrix.config.os == 'ubuntu-20.04'  
         shell: Rscript {0}
