+++ R/load_interface_terms.R
@@ -503,7 +503,7 @@
   # : en ----
   guidelines.en <- "
 <h1 id=archeoviz>archeoViz</h1>
-<p><code>archeoViz</code> is a packaged R Shiny application for the <em>visualisation</em>, <em>exploration</em>, and web <em>communication</em> of archaeological spatial data. It includes interactive 3D and 2D <em>visualisations</em>, can generate <em>cross sections</em> and <em>maps</em> of the remains, can run basic <em>spatial statistics</em> methods (convex hull, regression surfaces, 2D kernel density estimation), and display an interactive <em>timeline</em> of an excavation. <code>archeoViz</code> can be used locally or deployed on a server, either by allowing the user to load data through the interface or by running the app with a specific data set. The app interface is available in English, French, German, Italian, Portuguese, Romanian, and Spanish.</p>
+<p><code>archeoViz</code> is a packaged R Shiny application for the <em>visualisation</em>, <em>exploration</em>, and web <em>communication</em> of archaeological spatial data. It includes interactive 3D and 2D <em>visualisations</em>, can generate <em>cross sections</em> and <em>maps</em> of the remains, and display an interactive <em>timeline</em> of the work made in a site. Simple <em>spatial statistics</em> can be performed (convex hull, regression surfaces, 2D kernel density estimation), as well as <em>exporting data</em> to other online applications for more complex methods. <code>archeoViz</code> can be used locally or deployed on a server, either by allowing the user to load data through the interface or by running the app with a specific data set. The app interface is available in English, French, German, Italian, Portuguese, Romanian, and Spanish. Website: <a href=https://archeoviz.hypotheses.org class=uri>https://archeoviz.hypotheses.org</a>.</p>
 <ul>
 <li><a href=#installation><strong>Installation</strong></a>
 <ul>
@@ -799,7 +799,7 @@
 <li>select, from the menu that appears, the subsets of points for which convex hulls are to be calculated,</li>
 <li>click on \\\\u201cValidate\\\\u201d.</li>
 </ol>
-<p>The convex hulls associated with each subsets (with at least 20 points) are displayed. The convex hulls are computed using the <a href=https://CRAN.R-project.org/package=cxhull><code>cxhull</code></a> package.</p>
+<p>Convex hulls associated with each subsets with at least 20 points are displayed. The convex hulls are computed using the <a href=https://CRAN.R-project.org/package=cxhull><code>cxhull</code></a> package.</p>
 <h3 id=2d-kernel-density>2D kernel density</h3>
 <p>In the \\\\u201cMap\\\\u201d tab, ticking the \\\\u201cCompute density\\\\u201d box and clicking on \\\\u201cValidate\\\\u201d generates a map with contour lines showing the points\\\\u2019 density. Density can be computed for all the points together or by layer (with at least 30 points). The 2D kernel density is computed with the <code>kde2d</code> function of the <a href=https://CRAN.R-project.org/package=MASS><code>MASS</code></a> package (through <a href=https://CRAN.R-project.org/package=ggplot2><code>ggplot2</code></a>).</p>
 <h2 id=reproducibility>Reproducibility</h2>
@@ -812,8 +812,11 @@
[…]
