+++ NEWS.md
@@ -1,5 +1,11 @@
+# archeoViz 1.3.3
+Released: 2023-xx-xx
+
+* Add an export function to the 'expor' application for correspondance analysis
+* Add an export function to the 'amado' application for matrix seriation
+
 # archeoViz 1.3.2
-Released: 2023-09-XX
+Released: 2023-09-17
 
 * In the 3D view, the subsets of points for which convex hulls are computed can be selected (using the graphic user interface or the 'hulls.class.values' parameter of the archeoViz() function).
 * Add linguistic support for Rumanian.

[…]
