### 2023-10-10 ebbc1d5
mention grid square dimensions on plots
