+++ R/load_interface_terms_utf8.R
@@ -1937,6 +1937,7 @@ Cf. este <a href=https://analytics.huma-num.fr/archeoviz/tai>ejemplo</a>.</p>
     linear = "Linéaire",
     planar = "Planaire",
     volume = "Volumétrique",
+    grid = "Carroyage",
     click.on.point = "Cliquez sur un point pour afficher ses informations.",
     notif.objects.ok = "Fichier des objets ok!",
     notif.objects.not.ok = "Erreur. Certains champs requis sont absents, vérifiez les données.",
@@ -2010,6 +2011,7 @@ Cf. este <a href=https://analytics.huma-num.fr/archeoviz/tai>ejemplo</a>.</p>
     linear = "Linear",
     planar = "Planar",
     volume = "Volume",
+    grid = "Grid",
     year = "year",
     variable = "Variable",
     selection = "Selection",
@@ -2092,6 +2094,7 @@ Cf. este <a href=https://analytics.huma-num.fr/archeoviz/tai>ejemplo</a>.</p>
     linear = "Lineare",
     planar = "Planare",
     volume = "Volumetrica",
+    grid = "Grade",
     click.on.point = "Clique sobre um ponto para obter mais informações",
     notif.objects.ok = "Arquivo de objetos ok!",
     notif.objects.not.ok = "Erro. Certos campos exigidos estão ausentes. Por favor, verifique o arquivo.",
@@ -2170,6 +2173,7 @@ Cf. este <a href=https://analytics.huma-num.fr/archeoviz/tai>ejemplo</a>.</p>
     linear = "Lineare",
     planar = "Planare",
     volume = "Volumetrica",
+    grid = "Griglia",
     click.on.point = "Fare clic su un punto per visualizzarne le informazioni.",
     notif.objects.ok = "File degli oggetto ok!",
     notif.objects.not.ok = "Errore. Mancano alcuni campi obbligatori, controllare il file.",
@@ -2243,6 +2247,7 @@ Cf. este <a href=https://analytics.huma-num.fr/archeoviz/tai>ejemplo</a>.</p>
     linear = "Linear",
     planar = "Planar",
     volume = "Volumen",
+    grid = "Quadrat",
     year = "Jahr",
     variable = "Variable",
     selection = "Selection",
@@ -2325,6 +2330,7 @@ Cf. este <a href=https://analytics.huma-num.fr/archeoviz/tai>ejemplo</a>.</p>
     linear = "Lineal",
     planar = "Planar",
     volume = "Volumétrica",
+    grid = "Cuadrícula",
     click.on.point = "Haga clic en un punto para obtener más información",
     notif.objects.ok = "Archivo de objetos ok!",
     notif.objects.not.ok = "Error.Faltan algunos campos obligatorios. Por favor revise el archivo.",
@@ -2332,7 +2338,7 @@ Cf. este <a href=https://analytics.huma-num.fr/archeoviz/tai>ejemplo</a>.</p>
     notif.tick.value = "Marque al menos un valor (en el menú del lado izquierdo).",
     notif.error.coords.type = "Los valores xmin, ymin o zmin no son numéricos y se han convertido (posiblemente con pérdida).",
     notif.error.identifier = "Algunos de los identificadores únicos de los objetos están duplicados.",
-    notif.warn.obj.removed = "Se han eliminado los objetos con coordenadas incompletas..",
+    notif.warn.obj.removed = "Se han eliminado los objetos con coordenadas incompletas.",
     notif.warn.uncertainty = "Muchos objetos con ubicación incierta: podría ralentizar la aplicación.",
     notif.warn.refits = "Muchos remontajes: podría ralentizar la aplicación."
   )
[…]
