### 2023-10-18 #5811 @bischrob
@oliviaguest My review complete. The paper and software are of excellent quality. There were just a couple of minor revisions with a reference and some typos I noted here https://github.com/sebastien-plutniak/archeoviz/pull/13.
