+++ paper.bib
@@ -56,7 +56,7 @@
   title={Digital Archaeology Field Recording in the 4th Dimension: ArchField C++ a 4D GIS for Digital Field Work}, 
   year={2015},
   volume={2},
-  pages={251-258},
+  pages={251--258},
   doi={10.1109/DigitalHeritage.2015.7419505},
   Langid                   = {american}
   }
@@ -73,12 +73,12 @@
 }
 
 
-@Manual{Plutniak2023archeoviz1.2.1,
+@Manual{Plutniak2023archeoviz1.3.2,
   Title                    = {archeoViz. Visualisation, Exploration, and Web Communication of Archaeological Spatial Data},
-  Author                   = {Plutniak, Sébastien and Araujo, Renata and Giardino, Sara and Laabs, Julian},
+  Author                   = {Plutniak, Sébastien and Araujo, Renata and Coltofean, Laura and Delsol, Nicolas and Giardino, Sara and Laabs, Julian},
   Year                     = {2023},
-  Doi                      = {10.5281/zenodo.8092474},
-  Note                     = {R package version 1.2.1},
+  Doi                      = {10.5281/zenodo.8353270},
+  Note                     = {R package version 1.3.2},
   Url                      = {https://CRAN.R-project.org/package=archeoViz},
   Langid                   = {american}
 }
@@ -158,12 +158,12 @@
 
 @Article{RoyerEtal2023,
   Title                    = {SEAHORS: Spatial Exploration of ArcHaeological Objects in R Shiny},
-  Author                   = {Royer, Aurélien  and Discamps, Emmanuel and Plutniak, Sébastien and Thomas, Marc},
+  Author                   = {Royer, Aurélien and Discamps, Emmanuel and Plutniak, Sébastien and Thomas, Marc},
   Year                     = {2023},
-  Doi                      = {10.5281/zenodo.7674698},
-  pubstate                   = {forthcoming},
-
-  Journal                  = {PCI Archaeology},
+  Doi                      = {10.24072/pcjournal.289},
+  Number                   = {e55},
+  Volume                   = {3},
+  Journal                  = {Peer Community Journal},
   Langid                   = {american}
 }
 @Manual{Rcoreteam2023,
@@ -206,6 +206,6 @@
   Year                     = {2015},
   Pagetotal                = {135},
   note = {Version 2.2},
-   url = {http://legacy.ariadne-infrastructure.eu/wp-content/uploads/2019/01/ARIADNE_D2.2_Second_report_on_users_needs.pdf},
+  url = {http://legacy.ariadne-infrastructure.eu/wp-content/uploads/2019/01/ARIADNE_D2.2_Second_report_on_users_needs.pdf},
   Langid                   = {american}
 }
