### 2023-10-30 16f2780
fix test after fct edit, export links
