### 2023-11-11 #5811 @sebastien-plutniak
Thank you very much @joeroe for your review and suggestion, I will make these change and update the file. I have a question about:

 >     * [ ]  The formatting of inline citations in the latest proof doesn't look right to me.

Sorry, but I do not see where there is an issue, can you be more specific please? (which line(s) number(s)? Thanks!
