### 2023-11-13 #5811 @joeroe
In https://raw.githubusercontent.com/openjournals/joss-papers/joss.05811/joss.05811/10.21105.joss.05811.pdf, on line 8, the first set of citations read:

> [...] communication of archaeological spatial data (Plutniak et al., 2023), (Plutniak, 2023).

Where I'd expect:

> [...] communication of archaeological spatial data (Plutniak et al., 2023, Plutniak, 2023).

All the citations are like this. But maybe it's just JOSS' style?
