+++ README-fr.md
@@ -1,935 +1,0 @@
-archeoViz
-================
-
-`archeoViz` est une application dédiée à l’archéologie. Elle permet de
-*visualiser*, d’*explorer* interactivement, et d’exposer et
-*communiquer* rapidement sur le web des données archéologiques
-spatialisées. Elle propose des *visualisations* en 3D et 2D, génère des
-*coupes* et des *cartes* du mobilier archéologique, permet de visualiser
-la *chronologie* des travaux sur un site. Il est possible d’y réaliser
-des *statistiques spatiales* simples (enveloppes convexes, surfaces de
-régression, estimation de densité par noyau en 2D) ainsi que
-d’*exporter des données* vers d’autres applications en ligne pour
-l’exécution de méthodes plus complexes. `archeoViz` peut être
-utilisée localement ou déployée sur un serveur, soit en chargeant des
-données via l’interface, soit en lançant l’application avec un jeu de
-donnée spécifique. L’interface est disponible en allemand, anglais,
-italien, français, portugais et roumain. Page web:
-<https://archeoviz.hypotheses.org>.
-
-  - [**Installation**](#installation)
-      - [Locale](#locale)
-      - [Distante](#distante)
-      - [Démonstration](#démonstration)
-  - [**Recommandations
-    communautaires**](#recommandations-communautaires)
-      - [Signaler un bug](#signaler-un-bug)
-      - [Soumettre une modification](#soumettre-une-modification)
-  - [**Utilisation**](#utilisation)
-      - [L’information spatiale dans
-        archeoViz](#linformation-spatiale-dans-archeoviz)
-          - [Points, localisation exacte : les objets
-            relevés](#points-localisation-exacte--les-objets-relevés)
-          - [Points, localisation vague : passes, seaux, tamis et
-            erreurs
-            d’enregistement](#points-localisation-vague--passes-seaux-tamis-et-erreurs-denregistement)
-          - [Lignes](#lignes)
-          - [Surfaces](#surfaces)
-      - [Remontages et mesures de
-        fabrique](#remontages-et-mesures-de-fabrique)
-          - [Remontages](#remontages)
-          - [Mesures de fabrique](#mesures-de-fabrique)  
-      - [Format de données](#formater-des-données)
-          - [Formater des données](#formater-des-données)
-          - [Tableau des objets](#tableau-des-objets)
-          - [Tableau des remontages](#tableau-des-remontages)
-          - [Tableau de la chronologie](#tableau-de-la-chronologie)
-      - [Charger des données](#charger-des-données)
-          - [Via l’interface de
-            l’application](#via-linterface-de-lapplication)
-          - [Par génération de données
-            aléatoires](#par-génération-de-données-aléatoires)
-          - [Via les paramètres de la fonction
-            R](#via-les-paramètres-de-la-fonction-r)
-          - [Via les paramètres d’une URL](#via-les-paramètres-dune-url)
-      - [Sous-sélection de données](#sous-sélection-de-données)
-          - [Par mode de localisation](#par-mode-de-localisation)
-          - [Par couche ou catégorie
-            d’objet](#par-couche-ou-catégorie-dobjet)
-      - [Visualisations interactives](#visualisations-interactives)
-          - [Généralités](#généralités)
-          - [Sorties graphiques](#sorties-graphiques)
-          - [Visualisation de l’incertitude
-            spatiale](#visualisation-de-l-incertitude%20spatiale)  
-      - [Statistiques spatiales](#statistiques-spatiales)
-          - [Surfaces de régression](#surfaces-de-régression)
[…]
