### 2023-11-15 #5811 @sebastien-plutniak
> In https://raw.githubusercontent.com/openjournals/joss-papers/joss.05811/joss.05811/10.21105.joss.05811.pdf, on line 8, the first set of citations read:
>
> > [...] communication of archaeological spatial data (Plutniak et al., 2023), (Plutniak, 2023).
>
> Where I'd expect:
>
> > [...] communication of archaeological spatial data (Plutniak et al., 2023, Plutniak, 2023).
>
> All the multiple citations are like this. But maybe it's just JOSS' style?

@joeroe  Yes, I agree, the formatting you suggest would be better, but it seems impossible in the JOSS style (I've checked the [doc](https://joss.readthedocs.io/en/latest/submitting.html#citations)).
Any suggestion @oliviaguest ?

Besides this formatting issue, I think the paper is ready for publication!
