### 2023-11-15 #5811 @sebastien-plutniak
> I've also completed my review, sorry for the delay. I didn't find any major issues and I think it the software is a great fit for JOSS. I'll certainly be looking into it as a way to publish my own site data in the future.
>
> Some minor notes:
>
>     * [ ]  Given that the software has already been translated into six languages, I thought I might find some instructions for translating it into others in the contribution guidelines. Something to consider.
>
>     * [ ]  The formatting of inline citations in the latest proof doesn't look right to me.

@joeroe  The guidelines now include a “Translation” section, thanks for the suggestion!
