### 2023-11-16 #5811 @joeroe
> @joeroe  Yes, I agree, the formatting you suggest would be better, but it seems impossible in the JOSS style (I've checked the [doc](https://joss.readthedocs.io/en/latest/submitting.html#citations)).

Have you tried e.g. `[@Plutniak2023archeoviz1.3.2, @Plutniak2023bspf]`?
