### 2023-11-16 #5811 @sebastien-plutniak
> > @joeroe  Yes, I agree, the formatting you suggest would be better, but it seems impossible in the JOSS style (I've checked the [doc](https://joss.readthedocs.io/en/latest/submitting.html#citations)).
>
> Have you tried e.g. `[@Plutniak2023archeoviz1.3.2; @Plutniak2023bspf]`?

Yes, I did ([here](https://github.com/openjournals/joss-reviews/issues/5811#issuecomment-1813275914), although the result is no longer available, since the pdf has been updated); the result is worst.
