### 2023-11-16 #5811 @sebastien-plutniak
> > @joeroe  Yes, I agree, the formatting you suggest would be better, but it seems impossible in the JOSS style (I've checked the [doc](https://joss.readthedocs.io/en/latest/submitting.html#citations)).
>
> Have you tried e.g. `[@Plutniak2023archeoviz1.3.2; @Plutniak2023bspf]`?

No! I tried  `[@Plutniak2023archeoviz1.3.2, @Plutniak2023bspf]` and the result was even worst. But with “;” it works well, thanks for the trick! Now it looks fine!
