### 2023-11-23 #5811 @sebastien-plutniak
@oliviaguest the paper is ready, I do not want to to do more change!
The code is already archived on Zenodo, DOI: [10.5281/zenodo.7460193](https://doi.org/10.5281/zenodo.7460193),
or do you want me to create a new release related to the JOSS submission?
