### 2023-11-23 #5811 @oliviaguest
@sebastien-plutniak it should be the version that was the result of the JOSS review process, does that make sense? Because the publication at JOSS is the result of the review at JOSS (not a previous, or of course future, version).
