+++ R/load_interface_terms.R
@@ -515,6 +515,7 @@
 <ul>
 <li><a href=#reporting-bugs>Reporting Bugs</a></li>
 <li><a href=#suggesting-changes>Suggesting Changes</a></li>
+<li><a href=#translation>Translation</a></li>
 </ul></li>
 <li><a href=#use><strong>Use</strong></a>
 <ul>
@@ -637,6 +638,8 @@
 <p>If you encounter a bug, please fill an <a href=https://github.com/sebastien-plutniak/archeoviz/issues>issue</a> with all the details needed to reproduce it.</p>
 <h2 id=suggesting-changes>Suggesting changes</h2>
 <p>Suggestions of changes to <code>archeoViz</code> are welcome. These requests may concern additional functions, changes to documentation, additional examples, new features, etc. They can be made by filling an <a href=https://github.com/sebastien-plutniak/archeoviz/issues>issue</a> and, even better, using pull requests and the <a href=https://docs.github.com/articles/about-pull-requests>GitHub Fork and Pull model</a>.</p>
+<h2 id=translation>Translation</h2>
+<p>In the development of <code>archeoViz</code>, particular attention is paid to multilingualism. The application interface is available in several languages and translations into additional languages are welcome. To do so, please edit this <a href=https://github.com/sebastien-plutniak/archeoviz/blob/main/R/load_interface_terms_utf8.R>file</a> and submit a pull request.</p>
 <h1 id=use>Use</h1>
 <p>Having archaeological remains from a given site, <code>archeoViz</code> is designed to lower the technical barriers to fulfill three objectives:</p>
 <ul>
@@ -813,9 +816,9 @@
 <p><code>archeoViz</code> was designed as one of the building blocks of a decentralised digital ecosystem for archaeological data and analysis. In this perspective, features and functions are distributed in multiple interconnected applications, rather than concentrated into few systems. Consequently, data can be exported and imported between <code>archeoViz</code> and other web-based applications. Note that, so far, the export functionalities are only available when using online <code>archeoViz</code> instances.</p>
 <h3 id=export-from-archeoviz>Export from archeoViz</h3>
 <p>Data can be exported to other online applications from <code>archeoViz</code> \\\\u201cStatistics\\\\u201d tab. Some exports are possible only for specific types of data or if a minimum number of values is satisfied.</p>
-<p><a href=https://analytics.huma-num.fr/Sebastien.Plutniak/archeofrag><em>archeofrag</em></a> is an R package and web application to assess and evaluate the distinctions betwen archaeological spatial units (e.g. layers) based on the analysis of refitting relationships between fragments of objects. The web version of the application includes methods to measure the cohesion and admixture of spatial units, and compare it to simulated data. If an instance of <code>archeoViz</code> is launched with <a href=#refittings>refitting data</a>, then this data can be analysed with <code>archeofrag</code>. See an example <a href=https://analytics.huma-num.fr/archeoviz/grotte16>here</a>.</p>
[…]
