+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: archeoViz
 Type: Package
 Title: Visualisation, Exploration, and Web Communication of Archaeological Spatial Data
 Version: 1.3.3
-Date: 2023-10-09
+Date: 2023-11-23
 Authors@R: c(person("Sebastien", "Plutniak", email = "sebastien.plutniak@posteo.net", role = c("aut", "cre"), comment = c(ORCID = "0000-0002-6674-3806")),
              person("Renata", "Araujo", email = "rennyaraujo@yahoo.com", role = "ctb", comment = c(ORCID = "0000-0001-8874-2858")),
              person("Laura", "Coltofean", email = "laura.coltofean@gmail.com", role = "ctb", comment = c(ORCID = "0000-0002-2233-3565")),

[…]
