### 2023-11-30 #5811 @oliviaguest
@sebastien-plutniak why is this the way it is? Did they not contribute to the JOSS version of the code? Do they not wish to be named here? What reason is there to exclude their name and potential input?
