### 2023-11-30 #5811 @sebastien-plutniak
> @sebastien-plutniak why is this the way it is? Did they not contribute to the JOSS version of the code? Do they not wish to be named here? What reason is there to exclude their name and potential input?

@oliviaguest Sorry, I have not been clear enough in my previous [answer](https://github.com/openjournals/joss-reviews/issues/5811#issuecomment-1830899102): these persons are not included in the *authors* list of the JOSS paper because  (according to the CRAN's role taxonomy) they are *contributors* and neither “authors” nor  “creators” of the package.
Indeed, these persons did not contribute to the *code* of  the software (but only to the interface textual components, that they translated into  Spanish, Romanian, etc.).
Do you think they should be included in the JOSS paper author list? (consequently, as “authors”) ?
