### 2023-12-01 #5811 @sebastien-plutniak
> @sebastien-plutniak I get you now. Very clear. Thanks. Make no changes. ☺️
>
> What else is left pending at the moment?

Nothing! I've checked all the points listed in your [previous post](https://github.com/openjournals/joss-reviews/issues/5811#issuecomment-1824285428). I think we're done!
