### 2023-12-06 #5811 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
