+++ paper.bib
@@ -1,5 +1,5 @@
 @Book{Sievert2020,
-  Title                    = {Interactive Web-Based Data Visualization with R, plotly, and shiny},
+  Title                    = {Interactive Web-Based Data Visualization with {R}, plotly, and shiny},
   Author                   = {Sievert, Carson},
   Year                     = {2020},
   ISBN                     = {978-1138331495},
@@ -94,7 +94,7 @@
 }
 
 @Article{CouilletEtal2022,
-  Title                    = {New Visual Analytics Tool and Spatial Statistics to Explore Archeological Data: The Case of the Paleolithic Sequence of La Roche-à-Pierrot, Saint-Césaire, France},
+  Title                    = {New Visual Analytics Tool and Spatial Statistics to Explore Archeological Data: The Case of the Paleolithic Sequence of {La Roche-à-Pierrot, Saint-Césaire, France}},
   Author                   = {Couillet, Armelle and Rougier, Hélène and Todisco, Dominique and Marot, Josserand and Gillet, Olivier and Crevecoeur, Isabelle},
   Year                     = {2022},
   Doi                      = {10.5334/jcaa.81},
@@ -109,7 +109,7 @@
 
  @Book{Wood2017,
     title = {Generalized Additive Models},
-    subtitle = {An Introduction with R},
+    subtitle = {An Introduction with {R}},
     doi = {10.1201/9781315370279},
     year = {2017},
     author = {Wood, Simon N.},
@@ -131,7 +131,7 @@
   }
   
 @Manual{VaidyanathanEtal2023,
-    title = {htmlwidgets: HTML Widgets for R},
+    title = {htmlwidgets: HTML Widgets for {R}},
     author = {Vaidyanathan, Ramnath and  Xie, Yihui and Allaire, JJ and Cheng, Joe and Sievert, Carson and Russell, Kenton},
     year = {2023},
     note = {R package version 1.6.2},
