+++ paper.bib
@@ -73,12 +73,12 @@
 }
 
 
-@Manual{Plutniak2023archeoviz1.3.2,
+@Manual{Plutniak2023archeoviz1.3.3,
   Title                    = {archeoViz. Visualisation, Exploration, and Web Communication of Archaeological Spatial Data},
   Author                   = {Plutniak, Sébastien and Araujo, Renata and Coltofean, Laura and Delsol, Nicolas and Giardino, Sara and Laabs, Julian},
   Year                     = {2023},
-  Doi                      = {10.5281/zenodo.8353270},
-  Note                     = {R package version 1.3.2},
+  Doi                      = {10.5281/zenodo.10200892},
+  Note                     = {R package version 1.3.3},
   Url                      = {https://CRAN.R-project.org/package=archeoViz},
   Langid                   = {american}
 }
@@ -122,7 +122,7 @@
   }
 
   @Manual{Laurent2023,
-    title = {cxhull: Convex Hull},
+    title = {{cxhull}: Convex Hull},
     author = {Laurent, Stéphane},
     year = {2023},
     note = {R package version 0.7.2},
@@ -131,7 +131,7 @@
   }
   
 @Manual{VaidyanathanEtal2023,
-    title = {htmlwidgets: HTML Widgets for {R}},
+    title = {{htmlwidgets}: {HTML} Widgets for {R}},
     author = {Vaidyanathan, Ramnath and  Xie, Yihui and Allaire, JJ and Cheng, Joe and Sievert, Carson and Russell, Kenton},
     year = {2023},
     note = {R package version 1.6.2},
@@ -157,7 +157,7 @@
 
 
 @Article{RoyerEtal2023,
-  Title                    = {SEAHORS: Spatial Exploration of ArcHaeological Objects in R Shiny},
+  Title                    = {SEAHORS: Spatial Exploration of ArcHaeological Objects in {R} Shiny},
   Author                   = {Royer, Aurélien and Discamps, Emmanuel and Plutniak, Sébastien and Thomas, Marc},
   Year                     = {2023},
   Doi                      = {10.24072/pcjournal.289},
@@ -178,7 +178,7 @@
 }
 
 @Manual{ChangEtal2020,
-  Title                    = {shiny: Web Application Framework for R},
+  Title                    = {{shiny}: Web Application Framework for {R}},
   Author                   = {Chang, Winston and Cheng, Joe and Allaire, J.J. and Xie, Yihui and McPherson, Jonathan},
   Year                     = {2020},
   Note                     = {R package version 1.4.0.2},
