### 2023-12-12 #5811 @oliviaguest
Huge thanks to the reviewers @bischrob, @nevrome, @joeroe! ✨  JOSS appreciates your work and effort. ✨ Also, big congratulations to the author,  @sebastien-plutniak! 🥳 🍾
